requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
attrs==26.1.0
backoff==2.2.1
certifi==2026.2.25
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
deep-sort-realtime==1.3.2
defusedxml==0.7.1
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
idna==3.11
inference-sdk==1.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.3.5
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.2
pillow==11.3.0
propcache==0.4.1
py-cpuinfo==9.0.0
pyDeprecate==0.5.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
scipy==1.17.1
setuptools==70.2.0
six==1.17.0
supervision==0.27.0.post2
sympy==1.14.0
torch==2.11.0+cpu
torchvision==0.26.0+cpu
tqdm==4.67.3
typing-inspect==0.9.0
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
